backend/requirements.txt
fastapi 
uvicorn
mysql-connector-python
pandas
python-dotenv
bcrypt
gunicorn


